Move methods in appropriate file and review exports
Change-type: patch
