+++ src/extra/AutoUI/models/helpers.ts
@@ -9,9 +9,9 @@ import type {
 	JSONSchema7Definition as JSONSchemaDefinition,
 } from 'json-schema';
 import { Dictionary } from '../../../common-types';
-import { findInObject } from '../utils';
 import get from 'lodash/get';
-import pick from 'lodash/pick';
+import { getPropertyScheme } from '../../../components/Filters/SchemaSieve';
+import { isJson } from '../../../utils';
 
 type Transformers<
 	T extends Dictionary<any>,
@@ -29,15 +29,6 @@ export const autoUIDefaultPermissions = {
 	create: [],
 	update: [],
 	delete: false,
-};
-
-export const isJson = (str: string) => {
-	try {
-		JSON.parse(str);
-	} catch (err) {
-		return false;
-	}
-	return true;
 };
 
 export const autoUIRunTransformers = <
@@ -99,12 +90,6 @@ export const autoUIAdaptRefScheme = (
 	return transformed;
 };
 
-export const getSchemaFormat = (schema: JSONSchema) => {
-	const property = getSubSchemaFromRefScheme(schema);
-	const format = property.format ?? schema.format;
-	return format;
-};
-
 // This transformation would happen elsewhere, and it wouldn't be part of AutoUI
 export const autoUIGetModelForCollection = <T>(
 	model: AutoUIRawModel<T>,
@@ -143,90 +128,4 @@ export const autoUIAddToSchema = (
 			},
 		},
 	};
-};
-
-export const getPropertyScheme = (
-	schemaValue: JSONSchema | JSONSchemaDefinition,
-) => {
-	if (typeof schemaValue === 'boolean' || !schemaValue.description) {
-		return null;
-	}
-	try {
-		const json = JSON.parse(schemaValue.description!);
-		return json['x-foreign-key-scheme'] ?? json['x-ref-scheme'];
-	} catch (err) {
-		return null;
-	}
-};
-
-export const convertRefSchemeToSchemaPath = (refScheme: string | undefined) => {
-	return !!refScheme
-		? refScheme
-				.split('.')
-				.join('.properties.')
-				.replace(/\[\d+\]/g, '.items')
-		: // TODO: This atm doesn't support ['my property']
-		  refScheme;
-};
-
-export const getSchemaTitle = (
-	jsonSchema: JSONSchema,
-	propertyKey: string,
-	refScheme?: string | undefined,
-) => {
-	if (!refScheme) {
-		return jsonSchema?.title || propertyKey;
-	}
-	return (
-		getSubSchemaFromRefScheme(jsonSchema, refScheme)?.title ??
-		jsonSchema.title ??
-		propertyKey
-	);
-};
-
-export const generateSchemaFromRefScheme = (
-	schema: JSONSchema,
-	parentProperty: string,
-	refScheme: string | undefined,
-): JSONSchema => {
-	const propertySchema =
-		(schema.properties?.[parentProperty] as JSONSchema) ?? schema;
-	if (!refScheme) {
-		return propertySchema;
-	}
-	const convertedRefScheme = propertySchema.items
-		? `items.properties.${convertRefSchemeToSchemaPath(refScheme)}`
-		: `properties.${convertRefSchemeToSchemaPath(refScheme)}`;
-	const typePaths: string[][] = [];
-	const ongoingIncrementalPath: string[] = [];
-	convertedRefScheme.split('.').forEach((key) => {
-		if (['properties', 'items'].includes(key)) {
-			typePaths.push([...ongoingIncrementalPath, 'type']);
-		}
-		ongoingIncrementalPath.push(key);
-	});
-	if (ongoingIncrementalPath.length) {
-		typePaths.push(ongoingIncrementalPath);
-	}
-	return {
-		...propertySchema,
-		description: JSON.stringify({ 'x-ref-scheme': [refScheme] }),
-		title:
-			(get(propertySchema, convertedRefScheme) as JSONSchema)?.title ??
-			propertySchema.title,
-		...pick(propertySchema, typePaths),
-	};
-};
-
-export const getSubSchemaFromRefScheme = (
-	schema: JSONSchema | JSONSchemaDefinition,
-	refScheme?: string,
-): JSONSchema => {
-	const referenceScheme = refScheme || getPropertyScheme(schema)?.[0];
-	const convertedRefScheme = convertRefSchemeToSchemaPath(referenceScheme);
-	if (!convertedRefScheme) {
-		return schema as JSONSchema;
-	}
-	const properties = findInObject(schema, 'properties');
-	return get(properties, convertedRefScheme);
 };
