+++ src/utils/index.ts
@@ -109,3 +109,56 @@ export const removeFromLocalStorage = (key: string) => {
 		console.error(err);
 	}
 };
+
+export const isJson = (str: string) => {
+	try {
+		JSON.parse(str);
+	} catch (err) {
+		return false;
+	}
+	return true;
+};
+
+export const findInObject = (obj: any, key: string): any => {
+	let result;
+	for (const property in obj) {
+		if (obj.hasOwnProperty(property)) {
+			if (property === key) {
+				return obj[key]; // returns the value
+			} else if (typeof obj[property] === 'object') {
+				// in case it is an object
+				result = findInObject(obj[property], key);
+
+				if (typeof result !== 'undefined') {
+					return result;
+				}
+			}
+		}
+	}
+};
+
+export const onClickOutOrEsc = (
+	wrapper: HTMLDivElement,
+	callback: () => void,
+) => {
+	const handleClickOutside = (event: MouseEvent) => {
+		if (!wrapper.contains(event.target as Node)) {
+			handler();
+		}
+	};
+	const handleEsc = (event: KeyboardEvent) => {
+		if (event.key === 'Escape') {
+			handler();
+		}
+	};
+	const handler = () => {
+		callback();
+		if (document.activeElement instanceof HTMLElement) {
+			document.activeElement.blur();
+		}
+		document.removeEventListener('mousedown', handleClickOutside);
+		document.removeEventListener('keydown', handleEsc);
+	};
+	document.addEventListener('mousedown', handleClickOutside);
+	document.addEventListener('keydown', handleEsc);
+};
